Generalized tickers and the sample tables and feedback loops now work.
Removed the incomplete "playback" table.
Removed channelwise threading. It needs a serious rework.
Created the BasicTicker template, and made FloatTicker and IntTicker
typedefs of it.
Restructured the Wave classes such that arbitrary length and not
necessarily periodic waves can be given as output.
Restructured the PeriodicSampleTable class such that its internal
buffer depends more on what the WaveBase instance it takes dictates.
Added a currently-unused fill function to FxBase.
Fixed a bug in which any feedback value for an effect above 0 would cut
off audio.